% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMap.r
\name{baseMap}
\alias{baseMap}
\title{Create the base ggplot map}
\usage{
baseMap(
  coord_lim = list(lat = c(44, 49.1), long = c(-126.15, -122.12)),
  bath = list(include = TRUE, res = 1, breaks = c(-100, -30), colour = "grey", size =
    0.5),
  style = 1,
  xlab = "Longitude",
  ylab = "Latitude",
  main = "",
  stateBorders = FALSE,
  countryBorders = FALSE,
  wc_zm = 8,
  legend = FALSE
)
}
\arguments{
\item{coord_lim}{a list of lat and long coordinates for the limits of the plot}

\item{bath}{a list of bathymetric coordinates for "lat" and "long", resolution, colour, and size}

\item{style}{(1 = vanilla, 2 = bathymetry)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{main}{main title for the plot}

\item{stateBorders}{a vector of state borders c('washington','oregon') *no caps}

\item{wc_zm}{zoom level for style 4 watercolor map}

\item{legend}{for raster maps, should we show the legend? logical (T/F)}
}
\value{
ggplot map object
}
\description{
baseMap
}
\examples{

}
