% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectDb.r
\name{connectDb}
\alias{connectDb}
\title{Read in information from database using}
\usage{
connectDb(uid = default, pwd = default, db = default)
}
\arguments{
\item{uid}{User ID}

\item{pwd}{Default passwoed}

\item{db}{Database name}
}
\value{
channel A channel that can be passed to other functions that allows you to query the database
}
\description{
RODBC
}
