% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCPUE.R
\name{getCPUE}
\alias{getCPUE}
\title{Read in information from database using}
\usage{
getCPUE(
  uid = uid,
  pwd = pwd,
  db = "NWFSC_OCEAN",
  species = data.frame(commonName = "Chinook salmon", Age_ClassByLength = "yearling"),
  studyType = c("Regular", "Maximize Catch"),
  goodHaul = TRUE,
  day = TRUE,
  includeRepeats = TRUE,
  averageRepeats = FALSE,
  logData = "natural"
)
}
\arguments{
\item{uid}{User ID}

\item{pwd}{Default password}

\item{db}{Database name}

\item{species}{a list of species (commonName and Age_ClassByLength)}

\item{studyType}{most likely 'Regular' or 'Maximize Catch'}

\item{goodHaul}{use data from good hauls only?}

\item{day}{use hauls from daytime only?}

\item{includeRepeats}{include cpue from repeat trawls?}

\item{averageRepeats}{average cpue from all repeat (including nonconsecutive) trawls? This will provide station level data, not haul level data.}

\item{logData}{use 'N', 'natural', or 'base10'}
}
\value{
A data.frame with the CPUE data from the selected species
}
\description{
getCPUE
}
