% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getERDAPP.r
\name{getERDAPP}
\alias{getERDAPP}
\title{Read in information from database using}
\usage{
getERDAPP(var = NULL)
}
\arguments{
\item{var}{ERDAPP variable name}
}
\value{
a data fraem with results from the query
}
\description{
getQuery
}
