% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getERDDAP.r
\name{getERDDAP}
\alias{getERDDAP}
\title{Get an ERDDAP file}
\usage{
getERDDAP(var = NULL)
}
\arguments{
\item{var=NULL}{Dataset ID for an erddap file (see 'https://oceanview.pfeg.noaa.gov/erddap/info/index.html?page=1&itemsPerPage=1000')}
}
\value{
ERDDAP data.frame
}
\description{
Get an ERDDAP file
}
\examples{
#if you don't know the name of an erddap fileset just type
#and you'll be prompted to choose a fileset
getERDDAP()
#you can either enter the row number that you're interested in 
#or you can type something like 'PDO', and it will search the list for the PDO filesets
#Alternatively,
#If you don't want to look through all of the file sets, you just type
x <- getERDDAP(var='cciea')
#and you'll be asked to choose a row from the listed cciea filesets
}
