% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuery.r
\name{getQuery}
\alias{getQuery}
\title{Read in information from database using}
\usage{
getQuery(uid = uid, pwd = pwd, myQry = NA)
}
\arguments{
\item{uid}{User ID}

\item{pwd}{Default password}

\item{myQry}{A string representing a syntactically correct sql query}
}
\value{
a data fraem with results from the query
}
\description{
getQuery
}
