% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.r
\name{getTable}
\alias{getTable}
\title{Read in information from database using}
\usage{
getTable(uid = uid, pwd = pwd, db = "NWFSC_OCEAN", tableName = NA)
}
\arguments{
\item{uid}{User ID}

\item{pwd}{Default password}

\item{db}{Database name}
}
\value{
table A table from the list of tables that have already been built.
}
\description{
getTable
}
