% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeiplot.r
\name{oeiplot}
\alias{oeiplot}
\title{Title}
\usage{
oeiplot(
  df = data.frame(x = 1:4, y = 1:4),
  ylab = "",
  xlab = "",
  multipane = FALSE,
  ncol = 1
)
}
\arguments{
\item{df}{data.frame with the following column headings: x, y, Index}

\item{multipane}{Whether you want multicolors or multiple panes}

\item{ncol}{number of columns in the multipanel plot}
}
\value{
ggplot
}
\description{
Title
}
\examples{
oeiplot
}
