% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\title{Title}
\usage{
plot(
  df = data.frame(x = 1:4, y = 1:4),
  ylab = "",
  xlab = "",
  multipane = FALSE
)
}
\arguments{
\item{df}{data.frame with the following column headings: x, y, Index}

\item{multipane}{Whether you want multicolors or multiple panes}
}
\value{
ggplot
}
\description{
Title
}
\examples{
plot
}
