% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODBC.R
\name{readRODBC}
\alias{readRODBC}
\title{Read in information from database using}
\usage{
readRODBC()
}
\arguments{
\item{x}{some parameter we are interested in}
}
\value{
A table of information
}
\description{
RODBC
}
