% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseMap.r
\name{baseMap}
\alias{baseMap}
\title{Create the base ggplot map}
\usage{
baseMap(
  coord_lim = list(lat = c(44, 49.1), long = c(-126.15, -122.12)),
  bath = list(lat = c(44, 49.1), long = c(-126.15, -122.12), res = 3, breaks = c(-100),
    colour = "grey", size = 0.5),
  style = 1,
  xlab = "Longitude",
  ylab = "Latitude",
  stateBorders = FALSE
)
}
\arguments{
\item{coord_lim}{a list of lat and long coordinates for the limits of the plot}

\item{bath}{a list of bathymetric coordinates for "lat" and "long", resolution, colour, and size}

\item{style}{(1 = vanilla)}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{stateBorders}{a vector of state borders c('washington','oregon') *no caps}
}
\value{
ggplot map object
}
\description{
baseMap
}
\examples{

}
