% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getERDDAP.r
\name{getERDDAP}
\alias{getERDDAP}
\title{Get an ERDDAP file
getERDAPP}
\usage{
getERDDAP(var = NULL)
}
\arguments{
\item{var=NULL}{Dataset ID for an erddap file (see 'https://oceanview.pfeg.noaa.gov/erddap/info/index.html?page=1&itemsPerPage=1000')}
}
\value{
ERDDAP data.frame
}
\description{
Get an ERDDAP file
getERDAPP
}
\examples{
x <- getERDDAP(var='cciea')
}
