% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuery.r
\name{getQuery}
\alias{getQuery}
\title{Read in information from database using}
\usage{
getQuery(uid = uid, pwd = pwd, db = db, qryName = NA)
}
\arguments{
\item{uid}{User ID}

\item{pwd}{Default password}

\item{db}{Database name}

\item{qryName}{The name of a pre-built query in the database}
}
\value{
query A query from the list of queries that have already been built.
}
\description{
getQuery
}
